use uint16 for indices
A single mesh may not have more than max GLushort vertices so this
can already be restricted in mesh building.
(otherwise the vertex index will overflow VboMesh::compile)

+++ core/src/gl/vboMesh.h
@@ -115,10 +115,10 @@ protected:
 
     template <typename T>
     void compile(std::vector<std::vector<T>>& _vertices,
-                 std::vector<std::vector<int>>& _indices) {
+                 std::vector<std::vector<uint16_t>>& _indices) {
 
         std::vector<std::vector<T>> vertices;
-        std::vector<std::vector<int>> indices;
+        std::vector<std::vector<uint16_t>> indices;
 
         // take over contents
         std::swap(_vertices, vertices);
